+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.28.0
+* Add `upcomingRetryDate` to Transaction
+* Add `remainingFileEvidenceStorage` to `Dispute`
+* Add `transaction_retried` webhook
+* Add `transfer type` to Transaction
+
 ## 6.27.0
 * Add Session Id to Customer Recommendations Payload